Moved and renamed a bunch of things.
