+++ Source/Patches/Optional/ThirdPersonPatch.cs
@@ -15,7 +15,7 @@ namespace PlayerDogModel_Plus.Source.Patches.Optional
         [HarmonyPrefix]
         static bool ThirdPersonUpdatePrefix(ref Camera ____camera)
         {
-            if (!Plugin.boundConfig.thirdPersonConfigOverride.Value) return true; // Skip if we're not overriding.
+            if (!Plugin.config.thirdPersonConfigOverride.Value) return true; // Skip if we're not overriding.
 
             PlayerControllerB playerController = LocalPlayer.GetController();
             Camera gameplayCamera = playerController.gameplayCamera;
@@ -26,10 +26,10 @@ namespace PlayerDogModel_Plus.Source.Patches.Optional
 
             // Rest of this is ripped directly from the original method, minus the custom configs.
             Vector3 forwardVector = gameplayCamera.transform.forward * -1f;
-            Vector3 rightVector = gameplayCamera.transform.TransformDirection(Vector3.right) * Plugin.boundConfig.thirdPersonRightOffset.Value;
-            Vector3 upVector = Vector3.up * Plugin.boundConfig.thirdPersonUpOffset.Value;
+            Vector3 rightVector = gameplayCamera.transform.TransformDirection(Vector3.right) * Plugin.config.thirdPersonRightOffset.Value;
+            Vector3 upVector = Vector3.up * Plugin.config.thirdPersonUpOffset.Value;
 
-            float distance = Plugin.boundConfig.thirdPersonDistance.Value;
+            float distance = Plugin.config.thirdPersonDistance.Value;
 
             ____camera.transform.position = gameplayCamera.transform.position + forwardVector * distance + rightVector + upVector;
             ____camera.transform.rotation = Quaternion.LookRotation(gameplayCamera.transform.forward);
@@ -41,7 +41,7 @@ namespace PlayerDogModel_Plus.Source.Patches.Optional
         [HarmonyPrefix]
         static bool ThirdPersonOrbitUpdatePrefix(ref Camera ____camera)
         {
-            if (!Plugin.boundConfig.thirdPersonConfigOverride.Value) return true; // Skip if we're not overriding.
+            if (!Plugin.config.thirdPersonConfigOverride.Value) return true; // Skip if we're not overriding.
 
             PlayerControllerB playerController = LocalPlayer.GetController();
             Camera gameplayCamera = playerController.gameplayCamera;
@@ -61,7 +61,7 @@ namespace PlayerDogModel_Plus.Source.Patches.Optional
             // Rest of this is ripped directly from the original method, minus the custom configs.
             ____camera.transform.Rotate(Vector3.right, lookVector.y * _3rdPerson.Plugin.OrbitSpeedEntry.Value * Time.deltaTime);
             ____camera.transform.RotateAround(playerController.gameplayCamera.transform.position, Vector3.up, lookVector.x * _3rdPerson.Plugin.OrbitSpeedEntry.Value * Time.deltaTime);
-            float distance = Plugin.boundConfig.thirdPersonDistance.Value;
+            float distance = Plugin.config.thirdPersonDistance.Value;
             ____camera.transform.position = playerController.gameplayCamera.transform.position - ____camera.transform.forward * distance;
 
             return false;
